frontend: Merge directory edit and select pages
